правка подсчета стоимости путешествия
